refactor: rename serverless_llm to sllm
Renamed `serverless_llm` to `sllm` (pronounced "slim") for brevity.
