+++ sllm/serverless_llm/cli/delete.py
@@ -20,7 +20,7 @@ from argparse import Namespace, _SubParsersAction
 
 import requests
 
-from serverless_llm.serve.logger import init_logger
+from sllm.serve.logger import init_logger
 
 logger = init_logger(__name__)
