+++ sllm/serverless_llm/cli/replay.py
@@ -23,8 +23,8 @@ from argparse import Namespace, _SubParsersAction
 
 from openai import AsyncOpenAI
 
-from serverless_llm.cli._cli_utils import read_config
-from serverless_llm.serve.logger import init_logger
+from sllm.cli._cli_utils import read_config
+from sllm.serve.logger import init_logger
 
 logger = init_logger(__name__)
