+++ sllm/serverless_llm/serve/app_lib.py
@@ -21,7 +21,7 @@ import ray
 import ray.exceptions
 from fastapi import FastAPI, HTTPException, Request
 
-from serverless_llm.serve.logger import init_logger
+from sllm.serve.logger import init_logger
 
 logger = init_logger(__name__)
