+++ sllm/serverless_llm/serve/backends/vllm_backend.py
@@ -29,7 +29,7 @@ import torch
 from vllm import AsyncEngineArgs, AsyncLLMEngine, RequestOutput, SamplingParams
 from vllm.inputs import TokensPrompt
 
-from serverless_llm.serve.backends.backend_utils import (
+from sllm.serve.backends.backend_utils import (
     BackendStatus,
     SllmBackend,
 )
@@ -140,7 +140,7 @@ class VllmBackend(SllmBackend):
             storage_path = os.getenv("STORAGE_PATH", "./models")
             model_path = os.path.join(storage_path, "vllm", model)
             filtered_engine_config["model"] = model_path
-            filtered_engine_config["load_format"] = "serverless_llm"
+            filtered_engine_config["load_format"] = "sllm"
 
         # NOTE: Automatic enable prefix cachinging
         filtered_engine_config["enable_prefix_caching"] = True
