+++ sllm/serverless_llm/serve/inference_instance.py
@@ -15,48 +15,33 @@
 #  see the license for the specific language governing permissions and         #
 #  limitations under the license.                                              #
 # ---------------------------------------------------------------------------- #
-import argparse
-import asyncio
 
-from serverless_llm.cli.delete import DeleteCommand
-from serverless_llm.cli.deploy import DeployCommand
-from serverless_llm.cli.encode import EncodeCommand
-from serverless_llm.cli.generate import GenerateCommand
-from serverless_llm.cli.replay import ReplayCommand
-from serverless_llm.cli.update import UpdateCommand
-from serverless_llm.serve.logger import init_logger
+import ray
+
+from sllm.serve.logger import init_logger
 
 logger = init_logger(__name__)
 
 
-def main():
-    parser = argparse.ArgumentParser(
-        "sllm-cli", usage="sllm-cli <command> [<args>]"
-    )
-    commands_parser = parser.add_subparsers(help="sllm-cli command helpers")
+@ray.remote
+def start_instance(instance_id, backend, backend_config, startup_config):
+    logger.info(f"Starting instance {instance_id} with backend {backend}")
+    if backend == "vllm":
+        from sllm.serve.backends import VllmBackend
 
-    # Register commands
-    DeployCommand.register_subcommand(commands_parser)
-    GenerateCommand.register_subcommand(commands_parser)
-    EncodeCommand.register_subcommand(commands_parser)
-    ReplayCommand.register_subcommand(commands_parser)
-    DeleteCommand.register_subcommand(commands_parser)
-    UpdateCommand.register_subcommand(commands_parser)
+        model_backend_cls = ray.remote(VllmBackend)
+    elif backend == "dummy":
+        from sllm.serve.backends import DummyBackend
 
-    # Let's go
-    args = parser.parse_args()
+        model_backend_cls = ray.remote(DummyBackend)
+    elif backend == "transformers":
+        from sllm.serve.backends import TransformersBackend
 
-    if not hasattr(args, "func"):
-        parser.print_help()
-        exit(1)
-
-    # Run
-    service = args.func(args)
-    if asyncio.iscoroutinefunction(service.run):
-        asyncio.run(service.run())
+        model_backend_cls = ray.remote(TransformersBackend)
     else:
-        service.run()
+        logger.error(f"Unknown backend: {backend}")
+        raise ValueError(f"Unknown backend: {backend}")
 
-
-if __name__ == "__main__":
-    main()
+    return model_backend_cls.options(name=instance_id, **startup_config).remote(
+        backend_config
+    )
