+++ sllm/serverless_llm/serve/routers/roundrobin_router.py
@@ -22,7 +22,7 @@ from typing import Dict, Optional
 
 import ray
 
-from serverless_llm.serve.inference_instance import start_instance
+from sllm.serve.inference_instance import start_instance
 
 from .router_utils import InstanceHandle, SllmRouter
