+++ sllm/serverless_llm/serve/schedulers/storage_aware_scheduler.py
@@ -21,7 +21,7 @@ from typing import Mapping, Optional
 
 import ray
 
-from serverless_llm.serve.logger import init_logger
+from sllm.serve.logger import init_logger
 
 from .fcfs_scheduler import FcfsScheduler
