+++ sllm/serverless_llm/serve/store_manager.py
@@ -24,12 +24,12 @@ from typing import List, Mapping, Optional
 import ray
 from serverless_llm_store.client import SllmStoreClient
 
-from serverless_llm.serve.logger import init_logger
-from serverless_llm.serve.model_downloader import (
+from sllm.serve.logger import init_logger
+from sllm.serve.model_downloader import (
     VllmModelDownloader,
     download_transformers_model,
 )
-from serverless_llm.serve.utils import get_worker_nodes
+from sllm.serve.utils import get_worker_nodes
 
 logger = init_logger(__name__)
